+++ src/DateAndTime.php
@@ -6,7 +6,7 @@ namespace wdmg\helpers;
  * Yii2 Date and Time helper
  *
  * @category        Helpers
- * @version         1.4.0
+ * @version         1.4.1
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-helpers
  * @copyright       Copyright (c) 2019 - 2020 W.D.M.Group, Ukraine