+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## 1.4.1 (2020-10-21)
+ * Added `is_ip()`, `is_email()`, `is_mac()`, `is_serialized()`, `is_json()`, `is_geo()`, `is_domain()`, 
+   `is_url()`, `is_regexp()` methods to `StringHelper`
+ 
 ## 1.4.0 (2020-10-04)
  * Added `KeyboardLayoutHelper`
  * Added `DbSchemaTrait` for migrations